Spam Filtering
Implemented Naïve Bayes classification to classify emails as spam or not
spam.
In the training phase, raining data is used to learn a function that
maps text (vector of features) into classes.
In the testing phase, given a set of features and the function, the most
likely class is found i.e., spam or not spam.
Applied enhancements like smoothing and treatment of common words to
improve accuracy to 98%.
Developed using Python and PyCharm IDE